Change function signature of Parser::replace() and replaceAll() to remove exception
Since it would be a bit bothersome to encounter a crash and stack
trace one hour into an editing session the exception has been removed
from this function. It now returns a String indicating success, if it is
null and failure by returning the line where the mistake is missing.
Parser::replaceAll() has been adapted to this and now returns a list of
all mistakes where the replacement process has failed.