Rework error-handling in the TypoTattler- and Checker-constructor
The introduction of the FileNotFoundException in the Checker-constructor
allows distinguishing if there was a general IO-problem or if there was
a problem with finding/accessing files provided by the constructor
parameter. This hopefully helps with troubleshooting.

Additionally the error-messages are now more specific.

Modularisation and elimination of some redundancy have been made in the
name of readability.