Move the go-to-line-interaction out of the mainloop and fix it
Since the user-representation of lines starts with 1 and the
implementation starts counting from 0, 1 has to be substracted from the
user-input. Furthermore the mistake the program jumps to could be
invalid, so it has to look around for the next mistake. But it cannot
just invoke TypoTattler::next() or it would skip a mistake (because the
method calls Parser::hasNext().